cloud_deployment/cloud_function/requirements.txt
# Function dependencies, for example:
# package>=version
kaggle
google-cloud-secret-manager
google-cloud-storage
requests-toolbelt==1.0.0
google-auth==2.19.1
google-cloud-pubsub==2.21.5